+++ src/potent/operations/git_commit.py
@@ -4,24 +4,35 @@ from typing import Literal, override
 from potent.operations._base import BaseConfig, BaseOperation, OperationResult
 
 
-class Config(BaseConfig):
-    message: str
-    """
-    Commit message, submitted as is.
-    """
-    allow_empty: bool = False
-    """
-    If true, allows commits without changed/added files.
-    """
-
-
 class GitCommit(BaseOperation):
     """
     Commits staged files in git.
     """
 
+    class OpConfig(BaseConfig):
+        message: str
+        """
+        Commit message, submitted as is.
+        """
+        allow_empty: bool = False
+        """
+        If true, allows commits without changed/added files.
+        """
+
     slug: Literal["git-commit"] = "git-commit"
-    config: Config
+    config: OpConfig
+
+    @property
+    @override
+    def summary(self) -> str:
+        LIMIT = 10
+        msg = (
+            f"{self.config.message[:LIMIT]}..."
+            if len(self.config.message) > LIMIT
+            else self.config.message
+        )
+
+        return f'git commit -m "{msg}"'
 
     @override
     def _run(self, directory: Path) -> OperationResult: